add no-cache strategy
