+++ src/index.js
@@ -1,2 +1,3 @@
 export * from './DataCacheStrategy';
+export * from './NoCacheStrategy';
 export * from './OutputCacheDecorator';
